Added the functionality to the restricted pages

+++ base/views.py
@@ -3,8 +3,10 @@ from .models import Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.http import HttpResponse
 
 # rooms = [
 #     {"id": 1, "name": "Let's learn Python!"},
@@ -60,8 +62,10 @@ def room(request, pk):
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="/login")
 def createRoom(request):
     form = RoomForm()
+
     if request.method == "POST":
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -71,9 +75,13 @@ def createRoom(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="/login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You're not allowed here!")
 
     # Now we need to process this data and update the given room
     if request.method == "POST":
@@ -85,8 +93,13 @@ def updateRoom(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="/login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You're not allowed here!")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
